updating trees to include inner and outer BBC ADC values for east and west
The old tree version only stored BBC ADC east inner information. For the pAu 2015 analysis, we are adding the inner and outer information for both East and West BBCs. They were added in a completely backwards compatible way, so some of the names are not as clear as they could be, unfortunately.